Fix the switch, sign-compare and unused-variable warnings
persistence/file.cpp is the one that mattered. `File::Type version;` was
declared uninitialised and neither switch that assigns it is exhaustive: the
inner one has no case for AC1002/AC1003/AC1004, AC12/AC14/AC150/AC210, MC00 or
AC1032, and the outer one has no LIBOPENCAD case unless LIBOPENCAD_ENABLED is
defined. Opening a file of any unmapped revision therefore returned an
indeterminate File::Type, which the caller uses to choose the save format.
Initialise it to the R12 the existing UNKNOWNV case already falls back to, and
give both switches a default so the value is always one we actually set.

The rest:

lcUI/lcadviewer.cpp: the three QOpenGLDebugMessage switches did not handle
Qt's filter sentinels (AnySource/AnyType/AnySeverity) or the Invalid*
enumerators, which cannot arrive on a real message. Added defaults; the
severity switch logs them at debug rather than dropping them.

lcUI/lcadviewer.cpp: QWheelEvent::pos() is deprecated in favour of position().
position() is a QPointF and zoom() takes integral device coordinates, so the
result is rounded back through toPoint().

listgui.cpp, menuitem.cpp, toolbarbutton.cpp, vertexarray.cpp: loops compared
`int i` against `.size()`. Indexed with size_t instead.

lckernel/cad/storage/entitycontainer.h: entitiesByMetaType walked every entity
with a loop whose body was entirely commented out, so it built a vector,
discarded it, and left the loop variable unused. Commented the loop out with
the body it belongs to; the function returned an empty container before this
change and still does.

unittest/ui/widgets/layers.h: passing NULL for Qt::KeyboardModifiers selected
the deprecated QFlags(Zero) constructor. Now Qt::NoModifier.

unittest/ui/gui/testmenuitem.cpp: one test named the MainWindow it constructs
and never used it.

Verified by rebuilding lcunittest: -Wswitch, -Wsign-compare,
-Wdeprecated-declarations and -Wunused-variable all report zero, no new errors,
and DxfExportTest, ToolbarGroupTest, ToolbarTabTest, entitytest and LayerOpsTest
pass, 23 of 23.

Not addressed: the ~117 -Wreorder warnings. They are GCC-only -- Clang does not
raise them under this project's flags -- so they cannot be verified from this
machine, and they deserve their own pass since an initialiser that reads
another member is the one case where the ordering is a real bug rather than a
cosmetic mismatch.

+++ lcUI/widgets/guiAPI/toolbarbutton.cpp
@@ -114,14 +114,14 @@ void ToolbarButton::removeCallback(const char* cb_name) {
 }
 
 void ToolbarButton::callbackCalled() {
-    for (int i = 0; i < callbacks.size(); i++) {
+    for (size_t i = 0; i < callbacks.size(); i++) {
         // Nil-safe: nil callback's invoke() is a graceful no-op returning Nil.
         callbacks[i].invoke();
     }
 }
 
 void ToolbarButton::callbackCalledToggle(bool enabled) {
-    for (int i = 0; i < callbacks.size(); i++) {
+    for (size_t i = 0; i < callbacks.size(); i++) {
         callbacks[i].call(enabled);
     }
 }
